move rolling_ball to skimage.restoration
