+++ benchmarks/benchmark_morphology.py
@@ -77,53 +77,3 @@ class Skeletonize3d(object):
     def peakmem_skeletonize_3d(self):
         self.skeletonize(self.image)
 
-
-class RollingBall(object):
-    """Benchmark Rolling Ball algorithm."""
-
-    timeout = 120
-
-    def time_rollingball(self, radius):
-        morphology.rolling_ball(data.coins(), radius=radius)
-    time_rollingball.params = [25, 50, 75, 100, 150, 200]
-    time_rollingball.param_names = ["radius"]
-
-    def peakmem_reference(self, *args):
-        """Provide reference for memory measurement with empty benchmark.
-
-        Peakmem benchmarks measure the maximum amount of RAM used by a
-        function. However, this maximum also includes the memory used
-        during the setup routine (as of asv 0.2.1; see [1]_).
-        Measuring an empty peakmem function might allow us to disambiguate
-        between the memory used by setup and the memory used by target (see
-        other ``peakmem_`` functions below).
-
-        References
-        ----------
-        .. [1]: https://asv.readthedocs.io/en/stable/writing_benchmarks.html#peak-memory
-        """
-        pass
-
-    def peakmem_rollingball(self, radius):
-        morphology.rolling_ball(data.coins(), radius=radius)
-    peakmem_rollingball.params = [25, 50, 75, 100, 150, 200]
-    peakmem_rollingball.param_names = ["radius"]
-
-    def time_rollingball_nan(self, radius):
-        image = data.coins().astype(np.float_)
-        pos = np.arange(np.min(image.shape))
-        image[pos, pos] = np.NaN
-        morphology.rolling_ball(image, radius=radius, nansafe=True)
-    time_rollingball_nan.params = [25, 50, 75, 100, 150, 200]
-    time_rollingball_nan.param_names = ["radius"]
-
-    def time_rollingball_ndim(self):
-        path = data.image_fetcher.fetch('data/cells.tif')
-        image = io.imread(path)
-        morphology.rolling_ellipsoid(
-            image, kernel_shape=(1, 100, 100), intensity_axis=100)
-
-    def time_rollingball_threads(self, threads):
-        morphology.rolling_ball(data.coins(), radius=100, num_threads=threads)
-    time_rollingball_threads.params = range(0, 9)
-    time_rollingball_threads.param_names = ["threads"]
