+++ skimage/morphology/__init__.py
@@ -15,7 +15,6 @@ from .max_tree import (max_tree, area_opening, area_closing,
                        diameter_opening, diameter_closing,
                        max_tree_local_maxima)
 from ._deprecated import watershed
-from .rolling_ball import rolling_ball, rolling_ellipsoid
 
 __all__ = ['binary_erosion',
            'binary_dilation',
@@ -59,5 +58,4 @@ __all__ = ['binary_erosion',
            'diameter_opening',
            'diameter_closing',
            'max_tree_local_maxima',
-           'rolling_ellipsoid',
-           'rolling_ball']
+           ]
