+++ skimage/restoration/_rolling_ball_cy.pyx
