+++ skimage/restoration/rolling_ball.py
