+++ skimage/restoration/setup.py
@@ -17,6 +17,7 @@ def configuration(parent_package='', top_path=None):
             '_unwrap_3d.pyx',
             '_denoise_cy.pyx',
             '_nl_means_denoising.pyx'], working_path=base_path)
+    cython(['_rolling_ball_cy.pyx'], working_path=base_path)
 
     config.add_extension('_unwrap_1d', sources=['_unwrap_1d.c'],
                          include_dirs=[get_numpy_include_dirs()])
@@ -32,8 +33,11 @@ def configuration(parent_package='', top_path=None):
                          sources=['_nl_means_denoising.c'],
                          include_dirs=[get_numpy_include_dirs(),
                                        '../_shared'])
+    config.add_extension('_rolling_ball_cy', sources=['_rolling_ball_cy.c'],
+                         include_dirs=[get_numpy_include_dirs()])
 
     return config
+
 
 if __name__ == '__main__':
     from numpy.distutils.core import setup
